remove AES128-* and AES256-* from permitted ciphers list
The full list of ciphers here are what envoy permits by default [1].

They rate an "A" on qualy's ssl labs, but the AES128-* and AES256-*
ciphers are marked as "weak". I don't think they're *that* weak, but
it's true they don't support perfect forward secrecy, so maybe it's
worth turning them off?

[1] https://www.envoyproxy.io/docs/envoy/v1.7.1/api-v2/api/v2/auth/cert.proto.html?highlight=ciphers